% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beats-package.R
\docType{package}
\name{beats-package}
\alias{beats}
\alias{beats-package}
\title{beats: Import and Process ECG Data}
\description{
Designed to work with UFI devices. Reads binary
    data files (UBE) and provides a Shiny app for finding and exporting
    heart beats.
}
\author{
\strong{Maintainer}: Max Czapanskiy \email{maxczapanskiy@gmail.com}

}
\keyword{internal}
