% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ecg_bw190918_62R}
\alias{ecg_bw190918_62R}
\title{ECG of a blue whale}
\format{A data frame with 234,001 rows and two variables:
\describe{
  \item{timestamp}{time of record, in POSIXct format}
  \item{ecg}{value from ECG recorder in engineering units, ranges 0-4095}
}}
\usage{
ecg_bw190918_62R
}
\description{
A dataset containing the ECG record of a blue whale, collected 2019-09-18
off the coast of Big Sur, CA under NMFS permit no. #####.
}
\keyword{datasets}
