% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_hr.R
\name{find_hr}
\alias{find_hr}
\title{Find heart rate from ECG data}
\usage{
find_hr(data)
}
\arguments{
\item{data}{A data.frame with columns timestamp (POSIXct) and ecg (numeric)}
}
\value{
A data.frame with columns
\itemize{
\item timestamp (POSIXct)
\item ecg (numeric)
\item period_s (numeric)
\item freq_hz (numeric)
\item freq_bpm (numeric)
}
}
\description{
\code{find_hr} is an interactive tool for processing ECG data to get heart
rate profiles. Use this tool to identify individual heart beats and gaps in
the ECG data.
}
