% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beats-package.R
\docType{package}
\name{beats-package}
\alias{beats}
\alias{beats-package}
\title{beats: Heartbeat Detection in Electrocardiogram Data}
\description{
Import and process electrocardiogram (ECG) data.
    Reads binary data files from UFI devices (.ube files) and provides a
    Shiny app for finding and exporting heart beats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FlukeAndFeather/beats}
  \item Report bugs at \url{https://github.com/FlukeAndFeather/beats/issues}
}

}
\author{
\strong{Maintainer}: Max Czapanskiy \email{maxczapanskiy@gmail.com} (0000-0002-6302-905X)

}
\keyword{internal}
